Revert "reorganize views"
This reverts commit 9cca2cc2c502369c22612fa908fc0b62af3ac563.
