--- web_api/core/views/__init__.py
@@ -1,89 +1,0 @@
-from django.http import HttpRequest, HttpResponse, JsonResponse
-from django.shortcuts import get_object_or_404
-
-from core import auth
-from core.models import Account, PullRequestActivity
-
-
-@auth.login_required
-def activity(request: HttpRequest, team_id: str) -> HttpResponse:
-    account = get_object_or_404(
-        Account.objects.filter(memberships__user=request.user), id=team_id
-    )
-    kodiak_activity_labels = []
-    kodiak_activity_approved = []
-    kodiak_activity_merged = []
-    kodiak_activity_updated = []
-
-    total_labels = []
-    total_opened = []
-    total_merged = []
-    total_closed = []
-    for day_activity in PullRequestActivity.objects.filter(
-        github_installation_id=account.github_installation_id
-    ).order_by("date"):
-        kodiak_activity_labels.append(day_activity.date)
-        kodiak_activity_approved.append(day_activity.kodiak_approved)
-        kodiak_activity_merged.append(day_activity.kodiak_merged)
-        kodiak_activity_updated.append(day_activity.kodiak_updated)
-        total_labels.append(day_activity.date)
-        total_opened.append(day_activity.total_opened)
-        total_merged.append(day_activity.total_merged)
-        total_closed.append(day_activity.total_closed)
-
-    return JsonResponse(
-        dict(
-            kodiakActivity=dict(
-                labels=kodiak_activity_labels,
-                datasets=dict(
-                    approved=kodiak_activity_approved,
-                    merged=kodiak_activity_merged,
-                    updated=kodiak_activity_updated,
-                ),
-            ),
-            pullRequestActivity=dict(
-                labels=total_labels,
-                datasets=dict(
-                    opened=total_opened, merged=total_merged, closed=total_closed
-                ),
-            ),
-        )
-    )
-
-
-@auth.login_required
-def current_account(request: HttpRequest, team_id: str) -> HttpResponse:
-    account = get_object_or_404(
-        Account.objects.filter(memberships__user=request.user), id=team_id
-    )
-    return JsonResponse(
-        dict(
-            user=dict(
-                id=request.user.id,
-                name=request.user.github_login,
-                profileImgUrl=request.user.profile_image(),
-            ),
-            org=dict(
-                id=account.id,
-                name=account.github_account_login,
-                profileImgUrl=account.profile_image(),
-            ),
-            accounts=[
-                dict(
-                    id=x.id,
-                    name=x.github_account_login,
-                    profileImgUrl=x.profile_image(),
-                )
-                for x in Account.objects.filter(memberships__user=request.user)
-            ],
-        )
-    )
-
-
-@auth.login_required
-def ping(request: HttpRequest) -> HttpResponse:
-    return JsonResponse({"ok": True})
-
-
-def debug_sentry(request: HttpRequest) -> HttpResponse:
-    return HttpResponse(1 / 0)
